added react router structure, where app renders header (our links) and main (where our switch resides, showing/hiding components based on user browser's url
