+++ src/components/Header.js
@@ -1,0 +1,16 @@
+import React from 'react'
+import { Link } from 'react-router-dom'
+
+
+const Header = () => (
+  <header>
+    <nav>
+      <ul>
+        <li><Link to='/index'>Home</Link></li>
+        <li><Link to='/issues'>Issues</Link></li>
+      </ul>
+    </nav>
+  </header>
+)
+
+export default Header
