+++ src/index.js
@@ -3,6 +3,17 @@ import ReactDOM from 'react-dom';
 import './index.css';
 import App from './App';
 import registerServiceWorker from './registerServiceWorker';
+import {
+  BrowserRouter as Router,
+  Route,
+  Link
+} from 'react-router-dom'
 
-ReactDOM.render(<App />, document.getElementById('root'));
+
+ReactDOM.render((
+  <Router>
+    <App />
+  </Router>
+
+  ), document.getElementById('root'));
 registerServiceWorker();
